Update to allow path pass through for GET and POST
