+++ app/plugins/router.js
@@ -1,4 +1,5 @@
 const routes = [].concat(
+  require('../routes'),
   require('../routes/healthy'),
   require('../routes/healthz'),
   require('../routes/organisation')
